Move connor pkg to internal dir
