+++ internal/connor/LICENSE
