+++ internal/connor/connor.go
