+++ internal/connor/errors.go
