+++ internal/connor/ge.go
@@ -4,28 +4,28 @@ import (
 	"time"
 
 	"github.com/sourcenetwork/defradb/client"
-	"github.com/sourcenetwork/defradb/connor/numbers"
+	"github.com/sourcenetwork/defradb/internal/connor/numbers"
 )
 
-// le does value comparisons to determine whether one
-// value is strictly less than another.
-func le(condition, data any) (bool, error) {
+// ge does value comparisons to determine whether one
+// value is strictly larger than or equal to another.
+func ge(condition, data any) (bool, error) {
 	if condition == nil {
-		// Only nil is less than or equal to nil
-		return data == nil, nil
+		// Everything is greater than or equal to nil
+		return true, nil
 	}
 
 	switch c := condition.(type) {
 	case time.Time:
 		switch d := data.(type) {
 		case time.Time:
-			return d.Before(c) || d.Equal(c), nil
+			return d.After(c) || d.Equal(c), nil
 		case string:
 			dt, err := time.Parse(time.RFC3339, d)
 			if err != nil {
 				return false, err
 			}
-			return dt.Before(c) || dt.Equal(c), nil
+			return dt.After(c) || dt.Equal(c), nil
 		default:
 			return false, client.NewErrUnhandledType("data", d)
 		}
@@ -34,18 +34,18 @@ func le(condition, data any) (bool, error) {
 		case float64:
 			switch dn := numbers.TryUpcast(data).(type) {
 			case float64:
-				return dn <= cn, nil
+				return dn >= cn, nil
 			case int64:
-				return float64(dn) <= cn, nil
+				return float64(dn) > cn, nil
 			}
 
 			return false, nil
 		case int64:
 			switch dn := numbers.TryUpcast(data).(type) {
 			case float64:
-				return dn <= float64(cn), nil
+				return dn >= float64(cn), nil
 			case int64:
-				return dn <= cn, nil
+				return dn >= cn, nil
 			}
 
 			return false, nil
