+++ internal/connor/gt.go
@@ -4,28 +4,27 @@ import (
 	"time"
 
 	"github.com/sourcenetwork/defradb/client"
-	"github.com/sourcenetwork/defradb/connor/numbers"
+	"github.com/sourcenetwork/defradb/internal/connor/numbers"
 )
 
-// lt does value comparisons to determine whether one
-// value is strictly less than another.
-func lt(condition, data any) (bool, error) {
+// gt does value comparisons to determine whether one
+// value is strictly larger than another.
+func gt(condition, data any) (bool, error) {
 	if condition == nil {
-		// Nothing is less than nil
-		return false, nil
+		return data != nil, nil
 	}
 
 	switch c := condition.(type) {
 	case time.Time:
 		switch d := data.(type) {
 		case time.Time:
-			return d.Before(c), nil
+			return d.After(c), nil
 		case string:
 			dt, err := time.Parse(time.RFC3339, d)
 			if err != nil {
 				return false, err
 			}
-			return dt.Before(c), nil
+			return dt.After(c), nil
 		default:
 			return false, client.NewErrUnhandledType("data", d)
 		}
@@ -34,18 +33,18 @@ func lt(condition, data any) (bool, error) {
 		case float64:
 			switch dn := numbers.TryUpcast(data).(type) {
 			case float64:
-				return dn < cn, nil
+				return dn > cn, nil
 			case int64:
-				return float64(dn) < cn, nil
+				return float64(dn) > cn, nil
 			}
 
 			return false, nil
 		case int64:
 			switch dn := numbers.TryUpcast(data).(type) {
 			case float64:
-				return dn < float64(cn), nil
+				return dn > float64(cn), nil
 			case int64:
-				return dn < cn, nil
+				return dn > cn, nil
 			}
 
 			return false, nil
