+++ internal/connor/ilike.go
