+++ internal/connor/ilike_test.go
