+++ internal/connor/le.go
@@ -4,27 +4,28 @@ import (
 	"time"
 
 	"github.com/sourcenetwork/defradb/client"
-	"github.com/sourcenetwork/defradb/connor/numbers"
+	"github.com/sourcenetwork/defradb/internal/connor/numbers"
 )
 
-// gt does value comparisons to determine whether one
-// value is strictly larger than another.
-func gt(condition, data any) (bool, error) {
+// le does value comparisons to determine whether one
+// value is strictly less than another.
+func le(condition, data any) (bool, error) {
 	if condition == nil {
-		return data != nil, nil
+		// Only nil is less than or equal to nil
+		return data == nil, nil
 	}
 
 	switch c := condition.(type) {
 	case time.Time:
 		switch d := data.(type) {
 		case time.Time:
-			return d.After(c), nil
+			return d.Before(c) || d.Equal(c), nil
 		case string:
 			dt, err := time.Parse(time.RFC3339, d)
 			if err != nil {
 				return false, err
 			}
-			return dt.After(c), nil
+			return dt.Before(c) || dt.Equal(c), nil
 		default:
 			return false, client.NewErrUnhandledType("data", d)
 		}
@@ -33,18 +34,18 @@ func gt(condition, data any) (bool, error) {
 		case float64:
 			switch dn := numbers.TryUpcast(data).(type) {
 			case float64:
-				return dn > cn, nil
+				return dn <= cn, nil
 			case int64:
-				return float64(dn) > cn, nil
+				return float64(dn) <= cn, nil
 			}
 
 			return false, nil
 		case int64:
 			switch dn := numbers.TryUpcast(data).(type) {
 			case float64:
-				return dn > float64(cn), nil
+				return dn <= float64(cn), nil
 			case int64:
-				return dn > cn, nil
+				return dn <= cn, nil
 			}
 
 			return false, nil
