+++ internal/connor/ne.go
