+++ internal/connor/nilike_test.go
