+++ internal/connor/nin.go
