+++ internal/connor/not.go
