+++ internal/connor/numbers/equality.go
