+++ internal/connor/or.go
