+++ internal/connor/time/equality.go
