+++ planner/filter/unwrap_relation.go
@@ -10,7 +10,7 @@
 package filter
 
 import (
-	"github.com/sourcenetwork/defradb/connor"
+	"github.com/sourcenetwork/defradb/internal/connor"
 	"github.com/sourcenetwork/defradb/planner/mapper"
 )
