+++ planner/filter/unwrap_relation_test.go
@@ -13,7 +13,7 @@ import (
 	"testing"
 
 	"github.com/sourcenetwork/defradb/client/request"
-	"github.com/sourcenetwork/defradb/connor"
+	"github.com/sourcenetwork/defradb/internal/connor"
 	"github.com/sourcenetwork/defradb/planner/mapper"
 
 	"github.com/stretchr/testify/assert"
