+++ planner/filter/util_test.go
@@ -14,8 +14,8 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/sourcenetwork/defradb/connor"
 	"github.com/sourcenetwork/defradb/core"
+	"github.com/sourcenetwork/defradb/internal/connor"
 	"github.com/sourcenetwork/defradb/planner/mapper"
 )
