+++ planner/mapper/mapper.go
@@ -19,8 +19,8 @@ import (
 
 	"github.com/sourcenetwork/defradb/client"
 	"github.com/sourcenetwork/defradb/client/request"
-	"github.com/sourcenetwork/defradb/connor"
 	"github.com/sourcenetwork/defradb/core"
+	"github.com/sourcenetwork/defradb/internal/connor"
 )
 
 const (
